Adopt Riverpod codegen and reorganize models
