+++ lib/providers/dashboard/models/dashboard_state.freezed.dart
