+++ lib/providers/dashboard/models/enums/dashboard_list_filter.dart
@@ -1,0 +1,7 @@
+/// Filter applied to dashboard lists (tasks, tickets, meetings).
+enum DashboardListFilter {
+  all,
+  tasks,
+  tickets,
+  meetings,
+}
