--- lib/providers/dashboard_provider.dart
@@ -1,2 +1,0 @@
-export 'dashboard/dashboard_controller.dart';
-export 'dashboard/dashboard_state.dart';
