+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   # State management
   provider: ^6.1.2
   flutter_riverpod: ^2.6.1
+  riverpod_annotation: ^2.6.1
   freezed_annotation: ^2.4.4
 
   # Local storage
@@ -88,6 +89,7 @@ dev_dependencies:
   flutter_lints: ^5.0.0
   build_runner: ^2.4.13
   freezed: ^2.5.7
+  riverpod_generator: ^2.6.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
